+++ apps/app/src/components/editor/hideMarkers.ts
